Revert "SpriteObject velocity to float"
This reverts commit 21efc9664820698521a0154841f4ddce4aa2aa4c.

+++ src/Engine/Objects/SpriteObject.h
@@ -35,7 +35,7 @@ struct SpriteObject {
 
     inline void spellSpriteStop() {
         timeSinceCreated = 0_ticks;
-        vVelocity = Vec3f(0, 0, 0);
+        vVelocity = Vec3i(0, 0, 0);
     }
 
     /**
@@ -69,7 +69,7 @@ struct SpriteObject {
     // unsigned __int16 uType;
     uint16_t uObjectDescID = 0; // Index into pObjectList->pObjects. Zero means free slot, can reuse.
     Vec3f vPosition;
-    Vec3f vVelocity;
+    Vec3i vVelocity;
     uint16_t uFacing = 0;
     uint16_t uSoundID = 0;
     SpriteAttributes uAttributes = 0;
